% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcor.R
\name{bcor}
\alias{bcor}
\title{Standardized Ball Correlation for Multivariate data}
\usage{
bcor(x, y, fast = FALSE)
}
\arguments{
\item{x}{a numeric matirx}

\item{y}{a numeric matirx}

\item{fast}{Fast version for univariate case}
}
\description{
Calculate Ball Correlation statistic for Multivariate data
}
\examples{
n <- 100
x <- rnorm(n)
y <- rnorm(n)
bcor(x, y)

}
