% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningFsurv.R
\name{bcor.surv}
\alias{bcor.surv}
\title{Ball Correlation in survival analysis}
\usage{
bcor.surv(time, status, x)
}
\arguments{
\item{time}{Time}

\item{status}{Status}

\item{x}{single explanatory variable}
}
\value{
Survival Ball Correlation
}
\description{
Calculate Ball Correlation Statistic in survival
}
\examples{
data(survdat)
bcor.surv(time = survdat[["time"]], status = survdat[["status"]], x = survdat[,3])

}
