% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcorsis-package.r, R/screeningFrc.R
\docType{package}
\name{bcorsis}
\alias{bcorsis}
\alias{bcorsis-package}
\alias{bcorsis}
\title{(Iterative) Ball Correlation Sure Independence Screening ((I)SIS)}
\usage{
bcorsis(Y, X, candidate = c("large"), method = "bcorsis", parms = list(d1
  = 5, d2 = 5, df = 3), R = 99, parallel = FALSE, ncore = 2)
}
\arguments{
\item{Y}{a numeric matirx included univariate or multivariate response variable}

\item{X}{a numeric matirx included explanatory variables}

\item{candidate}{size of candidate set, only available to iterative BCor-SIS(I-BCor-SIS).
if candidate = "large", n variables are selected where n equal to sample size.
if candidate = "small", n/log(n) variables are selected.
if candidata is integer, candidata variables are selected.}

\item{method}{Method for screening procedure, include: "bcorsis", "bcorsis-pvalue",
"ibcorsis-lm", "ibcorsis-gam" and "interaction".
"bcorsis" means Standard Sure Independence Screening Procedure;
Another way for bcorsis procedure with p-value, is "bcorsis-pvalue".
"ibcorsis-lm" and "ibcorsis-gam" carry out Iterative BCor-SIS procedure with ordinary linear
regression and generalized additive models, respectively.
Option "interaction" is design for detect Variables with potential Interaction.}

\item{parms}{Parameters list only available for Iterative BCor-SIS. d1 is the
size of initial set, d2 is the variable set size added in each iteration.
df is degree freedom of basis in generalized additive models, so it only
play a role when method = "ibcorsis-gam".}

\item{R}{the number of replication. A argument only available when method == "bcorsis-pvalue"}

\item{parallel}{If parallel = TRUE, snowfall parallel framework will be used in features screening.}

\item{ncore}{If ncore = x, we use x threads to screening features.}
}
\description{
This function implements the (Iterative) Ball Correlation Sure Independence Screening for different variants of (I)SIS.

Ball Correlation Sure Independence Screening For Regression and Classification task(Include: BcorSIS, BCor-IISIS, Bcor-Interaction, Bcor-pvalue method)
}
\details{
bcorsis
}
\examples{
###############  Standard Method  ###############
data(regdat)
X <- regdat$X
Y <- regdat$Y
re <- bcorsis(X = X, Y = Y, candidate = "large")
re$candidate.set
###############  Iteration Method  ###############
data(interactiondat)
X <- interactiondat$X
Y <- interactiondat$Y
re<- bcorsis(X = X, Y = Y, candidate = "large", method = "interaction")
re$candidate.set
###############  BCorSIS procedure with p-value  ###############
data(regdat)
X <- regdat$X
Y <- regdat$Y
re <- bcorsis(X = X, Y = Y, candidate = "large", method = "bcorsis-pvalue", R = 10)
re$candidate.set
###############  Iterative Method  ###############
data(regdat)
re <- bcorsis(X = X, Y = Y, candidate = "large",
            method = "ibcorsis-gam", parms = list(d1 = 5, d2 = 5, df = 4))
re$candidate.set

}
\author{
WenLiang Pan, XueQin Wang, ChengFeng Liu, Jin Zhu
}
