% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningFsurv.R
\name{bcorsis.surv}
\alias{bcorsis.surv}
\title{Ball Correlation Sure Independence Screening For Survival data}
\usage{
bcorsis.surv(Y, X, candidate = "small", standized = TRUE)
}
\arguments{
\item{Y}{a numeric matirx(first column should be event time, second column should be survival status) or Surv object}

\item{X}{a numeric matirx included explanatory variables}

\item{candidate}{size of candidate set, only available to iterative BCor-SIS(I-BCor-SIS).
if candidate = "large", n variables are selected where n equal to sample size.
if candidate = "small", n/log(n) variables are selected.
if candidata is integer, candidata variables are selected.}

\item{standized}{allows the user to standardize the covariate}
}
\description{
Utilize extension of Ball Correlation in survival to select candidate variables related to survival status.
}
\examples{
data(survdat)
re <- bcorsis.surv(Y = survdat[, c(1, 2)], X = survdat[, -c(1, 2)], candidate = "large")
re$candidate.set

}
