% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{regdat}
\alias{regdat}
\title{Regression data for demonstration}
\format{A List contain Y and X. Sample sizes are 100 and variables are 100.}
\usage{
regdat
}
\description{
Y: response variable; X: covariate
}
\keyword{datasets}
