% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{survdat}
\alias{survdat}
\title{Survival data for demonstration}
\format{A data frame with 150 rows and 2500 variables}
\usage{
survdat
}
\description{
time: event time; status: event status; other: covariate
}
\keyword{datasets}
